Corrigindo erros de mapeamento em SistemaParametros e Rota

+++ src/main/java/br/gov/model/cadastro/SistemaParametros.java
@@ -1,4 +1,5 @@
 package br.gov.model.cadastro;
+import java.io.Serializable;
 import java.math.BigDecimal;
 
 import javax.persistence.Column;
@@ -8,8 +9,6 @@ import javax.persistence.Table;
 
 import br.gov.model.Status;
 import br.gov.model.util.ConstantesSistema;
-
-import java.io.Serializable;
 
 @Entity
 @Table(name="sistema_parametros", schema="cadastro")
@@ -99,7 +98,7 @@ public class SistemaParametros implements Serializable {
 	private String versaoCelular;
 
 	@Column(name="parm_icbloqueiocontamobile")
-	private Short indicadorBloqueioContaMobile;
+	private Integer indicadorBloqueioContaMobile;
 
 	@Column(name="parm_nndiascelular")
 	private Integer numeroDiasBloqueioCelular;
@@ -343,11 +342,11 @@ public class SistemaParametros implements Serializable {
 		this.versaoCelular = versaoCelular;
 	}
 
-	public Short getIndicadorBloqueioContaMobile() {
+	public Integer getIndicadorBloqueioContaMobile() {
 		return indicadorBloqueioContaMobile;
 	}
 
-	public void setIndicadorBloqueioContaMobile(Short indicadorBloqueioContaMobile) {
+	public void setIndicadorBloqueioContaMobile(Integer indicadorBloqueioContaMobile) {
 		this.indicadorBloqueioContaMobile = indicadorBloqueioContaMobile;
 	}
